% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{data.contention}
\alias{data.contention}
\title{data.contention}
\usage{
data.contention(data, sol)
}
\arguments{
\item{data}{data of asa problem}

\item{sol}{solution of asa problem}
}
\value{
data
}
\description{
data.contention
}
