% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{data.num.aircraft}
\alias{data.num.aircraft}
\title{Get the number of resources per period.}
\usage{
data.num.aircraft(sol)
}
\arguments{
\item{sol}{solution of WildfireResources problem.}
}
\value{
the number of aircraft in each period.
}
\description{
Get the number of resources per period.
}
