% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{data.performance}
\alias{data.performance}
\title{Performance in each period.}
\usage{
data.performance(data, sol)
}
\arguments{
\item{data}{data of asa problem.}

\item{sol}{solution of asa problem.}
}
\description{
Performance in each period.
}
