% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{data_num_resources}
\alias{data_num_resources}
\title{Get the number of resources per period.}
\usage{
data_num_resources(sol)
}
\arguments{
\item{sol}{solution of WildfireResources problem.}
}
\value{
the number of resources in each period.
}
\description{
Get the number of resources per period.
}
