% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(resources, fire, period_time)
}
\arguments{
\item{resources}{}

\item{fire}{}

\item{period_time}{}
}
\description{
Get data
}
\examples{
resources_file <- 'example/feasible/Resources.csv'
fire_file <- 'example/feasible/Fire.csv'
data <- load_data(resources_file, fire_file); data
resources <- data$data.resources
fire <- data$data.fire
period_time <- 10
}
