% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_model.R
\name{model}
\alias{model}
\title{Model}
\usage{
model(data)
}
\arguments{
\item{data}{data information.}
}
\description{
Model
}
\examples{
data <- WildfireResources::example_data()
m <- WildfireResources::model(data)
}
