% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{plot_num_resources}
\alias{plot_num_resources}
\title{Plot the number of resources per period.}
\usage{
plot_num_resources(df)
}
\arguments{
\item{df}{}
}
\value{
resources plot
}
\description{
Plot the number of resources per period.
}
