% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{plotnumaircraft}
\alias{plotnumaircraft}
\title{Plot the number of resources per period.}
\usage{
plotnumaircraft(df)
}
\arguments{
\item{df}{}
}
\value{
resources plot
}
\description{
Plot the number of resources per period.
}
