% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{plotscheduling}
\alias{plotscheduling}
\title{Scheduling plot}
\usage{
plotscheduling(WRF)
}
\arguments{
\item{WRF}{work, rest and fly matrix.}
}
\value{
scheduling plot.
}
\description{
Scheduling plot
}
