% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire_resources.R
\name{wildfire_resources}
\alias{wildfire_resources}
\title{Model to solve the Aircraft Selection and Allocation problem (ASA) for the containment of a wildfire.}
\usage{
wildfire_resources(data, solver = "gurobi", solver_params = list(TimeLimit =
  600, OutputFlag = 0))
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}

\item{solver_params}{list of gurobi options. Defaults to list(TimeLimit=600, OutputFlag = 0).}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Model to solve the Aircraft Selection and Allocation problem (ASA) for the containment of a wildfire.
}
\examples{
data <- WildfireResources::example_data()
WildfireResources::wildfire_resources(data, solver="lpSolveAPI")
}
