% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Wildfire Resources data.}
\usage{
load_data(resources_file, fire_file)
}
\arguments{
\item{resources_file}{}

\item{fire_file}{}
}
\description{
Load Wildfire Resources data.
}
